+++ include/NvInferVersion.h
@@ -19,9 +19,9 @@
 #define NV_INFER_VERSION_H
 
 #define NV_TENSORRT_MAJOR 9 //!< TensorRT major version.
-#define NV_TENSORRT_MINOR 1 //!< TensorRT minor version.
+#define NV_TENSORRT_MINOR 2 //!< TensorRT minor version.
 #define NV_TENSORRT_PATCH 0 //!< TensorRT patch version.
-#define NV_TENSORRT_BUILD 4 //!< TensorRT build number.
+#define NV_TENSORRT_BUILD 5 //!< TensorRT build number.
 
 #define NV_TENSORRT_LWS_MAJOR 0 //!< TensorRT LWS major version.
 #define NV_TENSORRT_LWS_MINOR 0 //!< TensorRT LWS minor version.
@@ -30,7 +30,7 @@
 // This #define is deprecated in TensorRT 8.6. Use NV_TENSORRT_MAJOR.
 #define NV_TENSORRT_SONAME_MAJOR 9 //!< Shared object library major version number.
 // This #define is deprecated in TensorRT 8.6. Use NV_TENSORRT_MINOR.
-#define NV_TENSORRT_SONAME_MINOR 1 //!< Shared object library minor version number.
+#define NV_TENSORRT_SONAME_MINOR 2 //!< Shared object library minor version number.
 // This #define is deprecated in TensorRT 8.6. Use NV_TENSORRT_PATCH.
 #define NV_TENSORRT_SONAME_PATCH 0 //!< Shared object library patch version number.