+++ plugin/modulatedDeformConvPlugin/modulatedDeformConvPluginKernel.h
@@ -31,10 +31,15 @@
 #include "modulatedDeformConvCudaHelper.h"
 #include <cstdint>
 #include <float.h>
+#include <cuda_fp16.h>
 
 template <typename T>
-__device__ T dmcnIm2colBilinear(
-    T const* input, int32_t const dataWidth, int32_t const height, int32_t const width, T h, T w);
+__device__ __forceinline__ T dmcnIm2colBilinear(
+    T const* input, int32_t const dataWidth, int32_t const height, int32_t const width, float h, float w);
+
+template <>
+__device__ __forceinline__ __half dmcnIm2colBilinear(
+    __half const* input, int32_t const dataWidth, int32_t const height, int32_t const width, float h, float w);
 
 template <typename T>
 __global__ void modulatedDeformableIm2colGpuKernel(int32_t const n, T const* dataIm, T const* dataOffset,
@@ -67,6 +72,12 @@ cudaError_t ModulatedDeformConvForwardCUDAKernelLauncher(TScalar const* input, T
 
 void ModulatedDeformConvForwardCUDAKernelLauncherFloat(float const* input, float const* weight, float const* bias,
     float const* offset, float const* mask, float* output, void* workspace, int32_t batch, int32_t channels,
+    int32_t height, int32_t width, int32_t channelsOut, int32_t kernelW, int32_t kernelH, int32_t strideW,
+    int32_t strideH, int32_t padW, int32_t padH, int32_t dilationW, int32_t dilationH, int32_t group,
+    int32_t deformableGroup, int32_t im2colStep, cublasHandle_t cublasHandle, cudaStream_t stream);
+
+void ModulatedDeformConvForwardCUDAKernelLauncherHalf(half const* input, half const* weight, half const* bias,
+    half const* offset, half const* mask, half* output, void* workspace, int32_t batch, int32_t channels,
     int32_t height, int32_t width, int32_t channelsOut, int32_t kernelW, int32_t kernelH, int32_t strideW,
     int32_t strideH, int32_t padW, int32_t padH, int32_t dilationW, int32_t dilationH, int32_t group,
     int32_t deformableGroup, int32_t im2colStep, cublasHandle_t cublasHandle, cudaStream_t stream);